tidying up code for minimap added extra files and started working on events
